+++ packages/raystack/components/data-view-beta/components/filters.tsx
@@ -114,6 +114,7 @@ export function Filters<TData>({
         label: field?.label || '',
         options: field?.filterOptions || [],
         selectProps: field?.filterProps?.select,
+        calendarProps: field?.filterProps?.calendar,
         ...filter
       };
     }) || [];
@@ -140,6 +141,7 @@ export function Filters<TData>({
               columnType={filter.filterType}
               options={filter.options}
               selectProps={filter.selectProps}
+              calendarProps={filter.calendarProps}
               className={classNames?.filterChips}
             />
           ))}